requirements.txt
aiogram
asyncpg
sqlalchemy
psycopg2-binary
python-dotenv 